Implement ignore paths
There are some paths we are not interested in, that only cause bloat.
Also for workloads such as databases (or anything else write-heavy) we
need a way of ignoring certain files. Hence it is now possible to ignore
paths via adding values to an eBPF map.

Additionally, the hashing algorithm has been switched to
`jenkins_one_at_a_time` to easy using the same hashing algorithm in eBPF
and user space.